src/01_behavioral_patterns/06_command_pattern/undoable_commands/commands/BoldCommand.ts
import CommandHistory from "../CommandHistory";
import HtmlDocument from "../HtmlDocument";
import ICommand from "../interfaces/ICommand";
import IUndoableCommand from "../interfaces/IUndoableCommand";

export default class BoldCommand implements ICommand, IUndoableCommand {
    private prevContent: string;
    private document: HtmlDocument;
    private history: CommandHistory;

    constructor(document: HtmlDocument, history: CommandHistory) {
        this.document = document;
        this.history = history;
    }

    public execute(): void {
        this.prevContent = this.document.getContent();
        this.document.makeBold();
        this.history.push(this);
    }
    
    public unexecute(): void {
        this.document.setContent(this.prevContent);
    }
}